requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
decorator==5.1.1
Django==3.2
django-celery-beat==2.5.0
django-cors-headers==3.13.0
django-filter==22.1
django-multiupload==0.6.1
django-timezone-field==5.0
djangorestframework==3.13.0
drf-yasg2==1.19.4
executing==1.2.0
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
kombu==5.2.4
magic-filter==1.0.9
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
multidict==6.0.4
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7
redis==4.5.2
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
